# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MYE030 Data Integration & Visualization API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/journals": {
      "get": {
        "tags": [
          "journals"
        ],
        "summary": "List Journals",
        "description": "Return a paginated list of journals, optionally filtered.\n\nAll free-text filters (``search_text``, ``publisher``,\n``best_subject_area``) are case-insensitive substring matches.\n``ranked_only=true`` hides rows without a quartile;\n``has_publisher`` and ``has_subject_area`` hide rows whose respective\ncolumn is NULL. Sortable columns: ``title``, ``publisher``,\n``best_quartile``, ``best_subject_area``, ``sjr_index``.",
        "operationId": "list_journals_api_journals_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          },
          {
            "name": "search_text",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search Text"
            }
          },
          {
            "name": "publisher",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Publisher"
            }
          },
          {
            "name": "best_quartile",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^Q[1-4]$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Best Quartile"
            }
          },
          {
            "name": "best_subject_area",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Best Subject Area"
            }
          },
          {
            "name": "ranked_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Ranked Only"
            }
          },
          {
            "name": "has_publisher",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Has Publisher"
            }
          },
          {
            "name": "has_subject_area",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Has Subject Area"
            }
          },
          {
            "name": "order_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Order By"
            }
          },
          {
            "name": "order_dir",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(asc|desc)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Order Dir"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_JournalSummary_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/journals/{journal_id}": {
      "get": {
        "tags": [
          "journals"
        ],
        "summary": "Get Journal Profile",
        "description": "Return the aggregated profile of one journal, optionally year-bounded.",
        "operationId": "get_journal_profile_api_journals__journal_id__get",
        "parameters": [
          {
            "name": "journal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Journal Id"
            }
          },
          {
            "name": "start_year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 2100,
                  "minimum": 1900
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Year"
            }
          },
          {
            "name": "end_year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 2100,
                  "minimum": 1900
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Year"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JournalProfile"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/journals/{journal_id}/articles/{article_id}": {
      "get": {
        "tags": [
          "journals"
        ],
        "summary": "Get Journal Article Details",
        "description": "Return full details of one journal article including its authors.",
        "operationId": "get_journal_article_details_api_journals__journal_id__articles__article_id__get",
        "parameters": [
          {
            "name": "journal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Journal Id"
            }
          },
          {
            "name": "article_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Article Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperDetails"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/journals/{journal_id}/yearly-statistics": {
      "get": {
        "tags": [
          "journals"
        ],
        "summary": "Get Journal Yearly Statistics",
        "description": "Return the per-year publication statistics that drive the line chart.\n\nTrims a trailing partial year by default; pass\n``include_partial_last_year=true`` to receive every row from the\nunderlying view.",
        "operationId": "get_journal_yearly_statistics_api_journals__journal_id__yearly_statistics_get",
        "parameters": [
          {
            "name": "journal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Journal Id"
            }
          },
          {
            "name": "start_year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 2100,
                  "minimum": 1900
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Year"
            }
          },
          {
            "name": "end_year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 2100,
                  "minimum": 1900
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Year"
            }
          },
          {
            "name": "include_partial_last_year",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Partial Last Year"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JournalYearlyStatistic"
                  },
                  "title": "Response Get Journal Yearly Statistics Api Journals  Journal Id  Yearly Statistics Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/journals/{journal_id}/articles": {
      "get": {
        "tags": [
          "journals"
        ],
        "summary": "List Journal Articles",
        "description": "Return a paginated list of articles published in the journal.\n\nSortable columns: ``title``, ``year``, ``pages``.",
        "operationId": "list_journal_articles_api_journals__journal_id__articles_get",
        "parameters": [
          {
            "name": "journal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Journal Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          },
          {
            "name": "start_year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 2100,
                  "minimum": 1900
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Year"
            }
          },
          {
            "name": "end_year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 2100,
                  "minimum": 1900
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Year"
            }
          },
          {
            "name": "order_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Order By"
            }
          },
          {
            "name": "order_dir",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(asc|desc)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Order Dir"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_JournalArticle_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/years/{year}/journals": {
      "get": {
        "tags": [
          "years"
        ],
        "summary": "List Year Journals",
        "description": "Return journals that published in the requested year.\n\nSortable columns: ``title``, ``publisher``, ``best_quartile``,\n``sjr_index``, ``articles_in_year``.",
        "operationId": "list_year_journals_api_years__year__journals_get",
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 2100,
              "minimum": 1900,
              "title": "Year"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          },
          {
            "name": "order_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Order By"
            }
          },
          {
            "name": "order_dir",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(asc|desc)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Order Dir"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_YearJournalEntry_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JournalProfile": {
        "properties": {
          "journal_id": {
            "type": "integer",
            "title": "Journal Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "publisher": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publisher"
          },
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "best_quartile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Quartile"
          },
          "best_subject_area": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Subject Area"
          },
          "sjr_index": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sjr Index"
          },
          "citation_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Citation Score"
          },
          "h_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "H Index"
          },
          "total_documents": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Documents"
          },
          "total_documents_3y": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Documents 3Y"
          },
          "total_references": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total References"
          },
          "total_citations_3y": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Citations 3Y"
          },
          "citable_documents_3y": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Citable Documents 3Y"
          },
          "citations_per_document_2y": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Citations Per Document 2Y"
          },
          "references_per_document": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "References Per Document"
          },
          "total_articles": {
            "type": "integer",
            "title": "Total Articles"
          },
          "earliest_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Earliest Year"
          },
          "latest_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Year"
          },
          "distinct_authors_total": {
            "type": "integer",
            "title": "Distinct Authors Total"
          },
          "average_articles_per_year": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Articles Per Year"
          },
          "average_authors_per_article_overall": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Authors Per Article Overall"
          }
        },
        "type": "object",
        "required": [
          "journal_id",
          "title",
          "total_articles",
          "distinct_authors_total"
        ],
        "title": "JournalProfile",
        "description": "Full profile combining ranking metrics and aggregate statistics."
      },
...
      "JournalYearlyStatistic": {
        "properties": {
          "year": {
            "type": "integer",
            "title": "Year"
          },
          "articles_count": {
            "type": "integer",
            "title": "Articles Count"
          },
          "distinct_authors": {
            "type": "integer",
            "title": "Distinct Authors"
          },
          "total_authors": {
            "type": "integer",
            "title": "Total Authors"
          },
          "average_authors_per_article": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Authors Per Article"
          }
        },
        "type": "object",
        "required": [
          "year",
          "articles_count",
          "distinct_authors",
          "total_authors"
        ],
        "title": "JournalYearlyStatistic",
        "description": "One bucket of the journal's per-year aggregate statistics."
      },
...
      "PaginatedResponse_JournalArticle_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/JournalArticle"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Page Size"
          },
          "total_items": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Items"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "page_size",
          "total_items"
        ],
        "title": "PaginatedResponse[JournalArticle]"
      },
      "PaginatedResponse_JournalSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/JournalSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Page Size"
          },
          "total_items": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Items"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "page_size",
          "total_items"
        ],
        "title": "PaginatedResponse[JournalSummary]"
      },
...
      "PaginatedResponse_YearJournalEntry_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/YearJournalEntry"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Page Size"
          },
          "total_items": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Items"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "page_size",
          "total_items"
        ],
        "title": "PaginatedResponse[YearJournalEntry]"
      },
...
      "PaperDetails": {
        "properties": {
          "article_id": {
            "type": "integer",
            "title": "Article Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          },
          "pages": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pages"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "venue_type": {
            "type": "string",
            "title": "Venue Type"
          },
          "venue_id": {
            "type": "integer",
            "title": "Venue Id"
          },
          "venue_title": {
            "type": "string",
            "title": "Venue Title"
          },
          "authors": {
            "items": {
              "$ref": "#/components/schemas/PaperAuthor"
            },
            "type": "array",
            "title": "Authors"
          }
        },
        "type": "object",
        "required": [
          "article_id",
          "title",
          "venue_type",
          "venue_id",
          "venue_title",
          "authors"
        ],
        "title": "PaperDetails",
        "description": "Full information about a single article including its authors."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}